+++ frontend/src/pages/edit-profile.tsx
@@ -13,7 +13,7 @@ import useTranslation from 'next-translate/useTranslation';
 
 import { GET_SELF_USER } from '@/lib/gql/User';
 
-const ProfileForm = dynamic(import('@/components/Web/Forms/profile'), { loading: () => <div /> });
+const ProfileForm = dynamic(import('@/user/forms/profile'), { loading: () => <div /> });
 const Layout = dynamic(import('@/layouts/Default'), { loading: () => <div /> });
 const Loader = dynamic(import('@/components/Web/Loader'), { loading: () => <div /> });
 const Error = dynamic(import('@/components/Common/Error'), { loading: () => <div /> });
