+++ frontend/src/pages/log-in.tsx
@@ -11,7 +11,7 @@ import useTranslation from 'next-translate/useTranslation';
 
 import { useAuth } from '@/contexts/AuthContext';
 import Layout from '@/layouts/Default';
-import LogInForm from '@/components/Web/Forms/log-in';
+import LogInForm from '@/user/forms/log-in';
 
 const SignIn = () => {
   const { t } = useTranslation('login');
