+++ frontend/src/pages/reset-password/[token].tsx
@@ -11,7 +11,7 @@ import { useRouter } from 'next/router';
 import useTranslation from 'next-translate/useTranslation';
 
 import { useAuth } from '@/contexts/AuthContext';
-import ResetPasswordForm from '@/components/Web/Forms/reset-password';
+import ResetPasswordForm from '@/user/forms/reset-password';
 import Layout from '@/layouts/Default';
 
 const ChangePassword = () => {
